Move trace loading to own module and class
