+++ miniml/gui/util.py
@@ -4,7 +4,6 @@ from importlib import resources
 import h5py
 
 import miniml.resources
-from miniml.core.trace import MiniTrace
 from miniml.fileio.util import is_keras_model
 
 
@@ -62,29 +61,3 @@ def get_hdf_keys(filepath: str) -> list:
     """
     with h5py.File(filepath, "r") as f:
         return list(f.keys())
-
-
-def load_trace_from_file(file_type: str, file_args: dict) -> MiniTrace:
-    """
-    Loads a trace from file and returns a MiniTrace object.
-
-    Parameters:
-    file_type (str): Type of file to load. Supported types are 'HEKA DAT', 'AXON ABF', and 'HDF5'.
-    file_args (dict): Dictionary of arguments to pass to the file loader.
-
-    Returns:
-    MiniTrace: MiniTrace object created from the loaded data.
-
-    Raises:
-    ValueError: If file_type is not supported.
-    """
-    file_loader = {
-        "HEKA DAT": MiniTrace.from_heka_file,
-        "AXON ABF": MiniTrace.from_axon_file,
-        "HDF5": MiniTrace.from_h5_file,
-    }.get(file_type, None)
-
-    if file_loader is None:
-        raise ValueError("Unsupported file type.")
-
-    return file_loader(**file_args)
